fix(firstbootwizard): ask root privilege to run fbw
Also, always try to login as root automatically before asking for password
