+++ plugins/lime-plugin-fbw/src/api.js
@@ -1,3 +1,4 @@
+import api from 'utils/uhttpd.service';
 
 export const searchNetworks = (api, rescan) =>
 	api.call('lime-fbw', 'search_networks', { scan: rescan || false });
@@ -7,3 +8,8 @@ export const setNetwork = (api, { file, hostname }) =>
 
 export const createNetwork = (api, { network, hostname, adminPassword }) =>
 	api.call('lime-fbw', 'create_network', { network, hostname, adminPassword });
+
+export const getFbwStatus = () =>
+	api.call('lime-fbw', 'status', {}).toPromise()
+		.catch(() => ({ lock: false }));
+
