+++ src/utils/routes.js
@@ -2,16 +2,16 @@ import { useAppContext } from './app.context';
 import { SharedPasswordLogin } from '../containers/SharedPasswordLogin';
 import { route } from 'preact-router';
 import { useEffect } from 'preact/hooks';
-import Fbw from '../../plugins/lime-plugin-fbw';
-import { useSession, useFbwStatus } from './queries';
-
+import { useSession } from './queries';
+import { useFbwStatus } from '../../plugins/lime-plugin-fbw/src/queries';
+import { FbwBanner } from '../../plugins/lime-plugin-fbw/src/containers/FbwBanner';
 
 export const Route = ({ path, children }) => {
 	const { data: fbwStatus } = useFbwStatus();
 	const { fbwCanceled } = useAppContext();
 
-	if (fbwStatus.lock && !fbwCanceled && path !== 'firmware' && path !== 'releaseInfo') {
-		return <Fbw.page />;
+	if (fbwStatus.lock && !fbwCanceled && path !== 'firmware' && path !== 'releaseInfo' && path !=='firstbootwizard') {
+		return <FbwBanner />
 	}
 
 	return children;
